add dummy consumer

+++ server/internal/pkg/trafficstream/type.go
@@ -2,5 +2,5 @@ package trafficstream
 
 //Message struct for holding traffic message
 type Message struct {
-	TrafficCount uint64
+	TrafficCount int
 }
